Remove landing page and make dashboard the home page
